Fixed net.h to solve mingw ws2tcpip.h incompatability errors
When compiling on mingw, when including compat.h as a sub-part of including checkpoints.h, ws2tcpip.h header is included. That header would complain that it is incompatible with winsock.h, meaning it was included somewhere before. It turned out that openssl/rand.h has a preprocessor case that includes windows.h, which inside winsock.h, causing the later error. Simply be moving the mruset/netbase/protocol/addrman includes above openssl/rand.h fixes the issue of comparability.