Cold minting
Users can generate a "cold minting" address from a "minting only" address and a "spending" address with the new RPC "addcoldmintingaddress".
Only a wallet having the minting key will be allowed to mint coins sent to this new address. And only a wallet with the spending key will be allowed to send the coins.

See src/test/features/cold_minting.feature for example scenarios.